refactor(pinsetstore::keychain): make keychain.rs private and reorganize/ consolidate keychain and pinset store tests
